chore: Atomise event helpers into individual files
