+++ common/helpers/events/get-container-classes.js
@@ -1,0 +1,7 @@
+const getEventClassification = require('./get-event-classification')
+
+const getContainerClasses = event => {
+  return getEventClassification(event) ? 'app-panel' : ''
+}
+
+module.exports = getContainerClasses
