+++ common/helpers/events/get-event-classification.js
@@ -1,0 +1,8 @@
+const getEventClassification = event => {
+  const { classification } = event
+  return classification && classification !== 'default'
+    ? classification
+    : undefined
+}
+
+module.exports = getEventClassification
