+++ common/helpers/events/get-flag.js
@@ -1,0 +1,18 @@
+const i18n = require('../../../config/i18n')
+
+const getEventClassification = require('./get-event-classification')
+
+const getFlag = event => {
+  const classification = getEventClassification(event)
+
+  if (!classification) {
+    return
+  }
+
+  return {
+    html: i18n.t(`events::classification.${classification}`),
+    type: classification,
+  }
+}
+
+module.exports = getFlag
