+++ common/helpers/events/index.js
@@ -1,0 +1,21 @@
+const addTriggeredEvents = require('./add-triggered-events')
+const getContainerClasses = require('./get-container-classes')
+const getDescription = require('./get-description')
+const getEventClassification = require('./get-event-classification')
+const getFlag = require('./get-flag')
+const getHeaderClasses = require('./get-header-classes')
+const getHeading = require('./get-heading')
+const getLabelClasses = require('./get-label-classes')
+const setEventDetails = require('./set-event-details')
+
+module.exports = {
+  addTriggeredEvents,
+  getContainerClasses,
+  getDescription,
+  getEventClassification,
+  getFlag,
+  getHeaderClasses,
+  getHeading,
+  getLabelClasses,
+  setEventDetails,
+}
