+++ common/presenters/move-to-timeline-component.js
@@ -1,9 +1,9 @@
-const addTriggeredEvents = require('../../common/helpers/events/add-triggered-events')
+const eventHelpers = require('../../common/helpers/events')
 
 const eventToTimelineItemComponent = require('./event-to-timeline-item-component')
 
 const moveToTimelineComponent = (move = {}) => {
-  move.timeline_events = addTriggeredEvents(move.timeline_events)
+  move.timeline_events = eventHelpers.addTriggeredEvents(move.timeline_events)
   const { timeline_events: timelineEvents = [] } = move
 
   const items = timelineEvents
