+++ common/presenters/move-to-timeline-component.test.js
@@ -4,7 +4,9 @@ const addTriggeredEvents = sinon.stub().returnsArg(0)
 const eventToTimelineItemComponent = sinon.stub().callsFake(val => val + val)
 
 const moveToTimelineComponent = proxyquire('./move-to-timeline-component', {
-  '../../common/helpers/events/add-triggered-events': addTriggeredEvents,
+  '../../common/helpers/events': {
+    addTriggeredEvents,
+  },
   './event-to-timeline-item-component': eventToTimelineItemComponent,
 })
